+++ mod/racing_plus/PostEntityKill.lua
@@ -435,7 +435,9 @@ function PostEntityKill:RoomClearDelayNPC(entity)
   -- The room clear delay NPC may accidentally die if Lua code kills all NPCs in a room
   -- If this occurs, just spawn a new one
   Isaac.DebugString("Room Clear Delay NPC death detected - spawning a new one.")
-  Isaac.Spawn(EntityType.ENTITY_ROOM_CLEAR_DELAY_NPC, 0, 0, g:GridToPos(0, 0), g.zeroVector, nil)
+  local roomClearDelayNPC = Isaac.Spawn(EntityType.ENTITY_ROOM_CLEAR_DELAY_NPC, 0, 0,
+                                        g:GridToPos(0, 0), g.zeroVector, nil)
+  roomClearDelayNPC:ClearEntityFlags(EntityFlag.FLAG_APPEAR) -- 1 << 2
 end
 
 -- After killing Mom, Mom's Heart, or It Lives!, all entities in the room are killed