+++ mod/CHANGES.md
@@ -102,6 +102,8 @@ While there are thousands of rooms in the game, many players have already seen t
 * Diagonal knife throws have a 3-frame window instead of a 1-frame window.
 * Duplicate rooms will no longer appear on the same run. (Basement 1 is exempt. All floors on set seeds are exempt.)
 * The Boss Rush is modified to include the Afterbirth+ bosses and Chapter 4 bosses.
+* Challenge Rooms are modified to include Womb enemies and bosses.
+* Boss Challenge Rooms will select random bosses instead of set bosses.
 * Maggy will automatically use her Speed Up pill at the beginning of the run.
 * Pin's first attack happens on the 15th frame (instead of the 73rd frame).
 * Cod Worms are replaced with Para-Bites.